Moved npis to src dir
